DotNet: Simplified 49.qnamaker-all-features.  README corrections.
