+++ samples/csharp_dotnetcore/49.qnamaker-all-features/Bots/QnABot.cs
@@ -8,6 +8,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Connector;
 using Microsoft.Bot.Schema;
 
 namespace Microsoft.BotBuilderSamples.Bots
@@ -27,6 +28,12 @@ namespace Microsoft.BotBuilderSamples.Bots
 
         public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default)
         {
+            // Teams group chat
+            if (turnContext.Activity.ChannelId.Equals(Channels.Msteams))
+            {
+                turnContext.Activity.Text = turnContext.Activity.RemoveRecipientMention();
+            }
+
             await base.OnTurnAsync(turnContext, cancellationToken);
 
             // Save any state changes that might have occurred during the turn.
