--- samples/csharp_dotnetcore/49.qnamaker-all-features/Dialog/RootDialog.cs
@@ -1,42 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Bot.Builder.AI.QnA.Dialogs;
-using Microsoft.Bot.Builder.Dialogs;
-
-namespace Microsoft.BotBuilderSamples.Dialog
-{
-    /// <summary>
-    /// This is an example root dialog. Replace this with your applications.
-    /// </summary>
-    public class RootDialog : ComponentDialog
-    {
-        /// <summary>
-        /// QnA Maker initial dialog
-        /// </summary>
-        private const string InitialDialog = "initial-dialog";
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RootDialog"/> class.
-        /// </summary>
-        /// <param name="services">Bot Services.</param>
-        public RootDialog(IBotServices services)
-            : base("root")
-        {
-            AddDialog(new QnAMakerBaseDialog(services));
-
-            AddDialog(new WaterfallDialog(InitialDialog)
-               .AddStep(InitialStepAsync));
-
-            // The initial child Dialog to run.
-            InitialDialogId = InitialDialog;
-        }
-
-        private async Task<DialogTurnResult> InitialStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            return await stepContext.BeginDialogAsync(nameof(QnAMakerDialog), null, cancellationToken);
-        }
-    }
-}
