+++ samples/csharp_dotnetcore/49.qnamaker-all-features/README.md
@@ -8,15 +8,32 @@ The [QnA Maker Service](https://www.qnamaker.ai) enables you to build, train and
 
 ## Concepts introduced in this sample
 The [QnA Maker Service][7] enables you to build, train and publish a simple question and answer bot based on FAQ URLs, structured documents or editorial content in minutes.
-In this sample, we demonstrate how to use the Active Learning to generate suggestions for knowledge base.
+In this sample, we demonstrate how to use the [Active Learning][8] to generate suggestions for knowledge base.
 
 # Prerequisites
-- Follow instructions [here](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/set-up-qnamaker-service-azure) to create a QnA Maker service.
-- Follow instructions [here](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/multiturn-conversation) to create multiturn experience.
-- Follow instructions [here](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/quickstarts/create-publish-knowledge-base) to import and publish your newly created QnA Maker service.
-- Update [appsettings.json](appsettings.json) with your kbid (KnowledgeBase Id), endpointKey and endpointHost. QnA knowledge base setup and application configuration steps can be found [here](https://aka.ms/qna-instructions).
-- (Optional) Follow instructions [here](https://github.com/Microsoft/botbuilder-tools/tree/master/packages/QnAMaker) to set up the
-QnA Maker CLI to deploy the model.
+
+This samples **requires** prerequisites in order to run.
+
+### Overview
+
+This bot uses [QnA Maker Service](https://www.qnamaker.ai), an AI based cognitive service, to implement simple Question and Answer conversational patterns.
+
+### Create a QnAMaker Application to enable QnA Knowledge Bases
+
+- QnA knowledge base setup and application configuration steps can be found [here](https://aka.ms/qna-instructions).
+  - Note that these instructions are for a different sample, but still apply to this sample.  Namely: Create a QnA service, and import 'cognitiveModels/smartLightFAQ.tsv' to a knowledgebase
+  - The knowledgebase 'Deployment Details' will show something like:
+    ```bash
+    POST /knowledgebases/04e394a2-7d8a-4a56-9344-db13fb1d9d79/generateAnswer
+    Host: https://servicename.azurewebsites.net/qnamaker
+    Authorization: EndpointKey bc3fbc8e-f461-66b1-8c58-4d6799320f9c
+    Content-Type: application/json
+    {"question":"<Your question>"}
+    ```
+      - QnAKnowledgebaseId is '04e394a2-7d8a-4a56-9344-db13fb1d9d79'
+      - QnAEndpointKey is 'bc3fbc8e-f461-66b1-8c58-4d6799320f9c'
+      - QnAEndpointHostName is 'https://servicename.azurewebsites.net/qnamaker'
+  - Enable [Active Learning][8] on your QnA service.
 
 ## To try this sample
 
@@ -27,6 +44,7 @@ QnA Maker CLI to deploy the model.
     ```
 
 - In a terminal, navigate to `samples/csharp_dotnetcore/49.qnamaker-all-features`
+- Update [appsettings.json](appsettings.json) with your QnAKnowledgebaseId, QnAEndpointKey and QnAEndpointHostName.
 - Run the bot from a terminal or from Visual Studio, choose option A or B.
 
   A) From a terminal
@@ -43,31 +61,6 @@ QnA Maker CLI to deploy the model.
   - Navigate to `samples/csharp_dotnetcore/49.qnamaker-all-features` folder
   - Select `QnABot.csproj` file
   - Press `F5` to run the project
-
-##### Microsoft Teams channel group chat fix
-- Goto `Bot/QnABot.cs`
-- Add References
-    ~~~
-    using Microsoft.Bot.Connector;
-    using System.Text.RegularExpressions;
-    ~~~
-- Modify `OnTurnAsync` function as:
-    ~~~
-    public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default)
-        {
-            // Teams group chat
-            if (turnContext.Activity.ChannelId.Equals(Channels.Msteams))
-            {
-                turnContext.Activity.Text = turnContext.Activity.RemoveRecipientMention();
-            }
-            
-            await base.OnTurnAsync(turnContext, cancellationToken);
-
-            // Save any state changes that might have occurred during the turn.
-            await ConversationState.SaveChangesAsync(turnContext, false, cancellationToken);
-            await UserState.SaveChangesAsync(turnContext, false, cancellationToken);
-        }
-    ~~~
 
 ## Testing the bot using Bot Framework Emulator
 
@@ -89,8 +82,10 @@ The deployment process assumes you have an account on Microsoft Azure and are ab
 If you are new to Microsoft Azure, please refer to [Getting started with Azure][70] for guidance on how to get started on Azure.
 
 # Further reading
-* [Active learning Documentation][al#1]
-* [Bot Framework Documentation][80]
+* [Manage QnA Maker resources](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/set-up-qnamaker-service-azure)
+* [Use follow-up prompts to create multiple turns of a conversation](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/multiturn-conversation)
+* [Quickstart: Create, train, and publish your QnA Maker knowledge base](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/quickstarts/create-publish-knowledge-base)
+* [Active learning Documentation][8]* [Bot Framework Documentation][80]
 * [Bot Basics][90]
 * [Azure Bot Service Introduction][100]
 * [Azure Bot Service Documentation][110]
@@ -104,7 +99,7 @@ If you are new to Microsoft Azure, please refer to [Getting started with Azure][
 [5]: https://github.com/microsoft/botframework-emulator
 [6]: https://aka.ms/botframeworkemulator
 [7]: https://www.qnamaker.ai
-
+[8]: https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/improve-knowledge-base
 [50]: https://docs.microsoft.com/en-us/azure/bot-service/bot-builder-howto-deploy-azure?view=azure-bot-service-4.0
 [60]: https://portal.azure.com
 [70]: https://azure.microsoft.com/get-started/
