+++ samples/csharp_dotnetcore/49.qnamaker-all-features/Startup.cs
@@ -30,9 +30,6 @@ namespace Microsoft.BotBuilderSamples
             // Create the Bot Framework Adapter with error handling enabled.
             services.AddSingleton<IBotFrameworkHttpAdapter, AdapterWithErrorHandler>();
 
-            // Create the bot services(QnA) as a singleton.
-            services.AddSingleton<IBotServices, BotServices>();
-
             // Create the storage we'll be using for User and Conversation state. (Memory is great for testing purposes.)
             services.AddSingleton<IStorage, MemoryStorage>();
 
@@ -43,10 +40,10 @@ namespace Microsoft.BotBuilderSamples
             services.AddSingleton<ConversationState>();
 
             // The Dialog that will be run by the bot.
-            services.AddSingleton<RootDialog>();
+            services.AddSingleton<BotQnAMakerDialog>();
 
             // Create the bot as a transient. In this case the ASP Controller is expecting an IBot.
-            services.AddTransient<IBot, QnABot<RootDialog>>();
+            services.AddTransient<IBot, QnABot<BotQnAMakerDialog>>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
